requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=10.0.0
matplotlib>=3.7.0
pandas>=2.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
psutil>=5.9.0
pytest>=7.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0